# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: User Management Control Panel API
  version: 0.1.0
  description: |
    Control-plane API for the user-management + licensing system.

    Covers admin authentication, organization/user/RBAC management, plan & subscription
    administration, license JWKS publication, key rotation, usage ingestion, audit log
    access, SSO configuration, API key management, and the domain event stream.

    Two authentication schemes are supported:
      - **bearerAuth**: short-lived admin session JWT, issued by `POST /api/v1/auth/login`.
      - **apiKeyAuth**: long-lived programmatic key, sent as `Authorization: ApiKey <token>`.

    The public `/.well-known/jwks.json` endpoint is unauthenticated.

    Error responses follow [RFC 7807 Problem Details](https://datatracker.ietf.org/doc/html/rfc7807).

    > **Caveat:** the license-issuance endpoint (`POST .../licenses`) referenced in the
    > implementation plan is *not yet wired up* in `control-panel-api`. It is documented
    > here as the intended contract for Phase 1 closure.
  contact:
    name: Control Panel Team
    email: support@control-panel.example.com
  license:
    name: Proprietary
servers:
  - url: https://control-panel.example.com
    description: Production
  - url: https://staging.control-panel.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/rbac/users/{userId}/roles:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    post:
      tags: [rbac]
      summary: Assign a role to a user (optionally scoped to an org).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '201':
          description: Assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleAssignment'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/rbac/users/{userId}/roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
      - in: path
        name: roleId
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: orgId
        required: false
        schema: { type: string, format: uuid }
    delete:
      tags: [rbac]
      summary: Remove a role assignment.
      responses:
        '204':
          description: Removed.
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # plans
  # ============================================================
  /api/v1/plans:
    get:
      tags: [plans]
      summary: List plans.
      parameters:
        - in: query
          name: activeOnly
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
    post:
      tags: [plans]
      summary: Create a plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /api/v1/plans/{planId}:
    parameters:
      - in: path
        name: planId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [plans]
      summary: Get a plan.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [plans]
      summary: Update top-level plan attributes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /api/v1/plans/{planId}/permissions:
    parameters:
      - in: path
        name: planId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [plans]
      summary: Replace the plan's permission set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionCodes]
              properties:
                permissionCodes:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /api/v1/plans/{planId}/features:
    parameters:
      - in: path
        name: planId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [plans]
      summary: Replace the plan's feature map.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [features]
              properties:
                features:
                  type: object
                  additionalProperties: true
                  description: |
                    Map of feature key -> value. Permitted value types: boolean, integer, string.
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'

  # ============================================================
  # subscriptions
  # ============================================================
...
  /api/v1/orgs/{orgId}/sso:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [sso]
      summary: List SSO providers configured for an org.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SsoProvider'
    post:
      tags: [sso]
      summary: Configure an SSO provider (SAML 2.0 or OIDC).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoProviderCreate'
      responses:
        '200':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsoProvider'
  /api/v1/orgs/{orgId}/sso/{providerId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - in: path
        name: providerId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [sso]
      summary: Remove an SSO provider config.
      responses:
        '204':
          description: Removed.
  /api/v1/orgs/{orgId}/sso/{providerId}/test:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - in: path
        name: providerId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [sso]
      summary: Dry-run a configured SSO provider.
      responses:
        '200':
          description: Test outcome.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SsoTestResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session JWT from `POST /api/v1/auth/login`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Long-lived programmatic key.
        Format: `Authorization: ApiKey <token>` where `<token>` is the value returned
        once by `POST /api/v1/orgs/{orgId}/api-keys`.
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string, format: uuid }
    UserIdParam:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Resource does not exist or is hidden from this caller.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflicting state (e.g. unique-constraint violation).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Bad Request' }
        status: { type: integer, example: 400 }
        detail: { type: string, example: 'Field `email` is required.' }
        instance: { type: string, format: uri }
        code: { type: string, example: 'validation_failed' }
      required: [title, status]
...
    AssignRoleRequest:
      type: object
      required: [roleCode]
      properties:
        roleCode: { type: string }
        orgId: { type: string, format: uuid, nullable: true }
    UserRoleAssignment:
      type: object
      properties:
        userId: { type: string, format: uuid }
        roleId: { type: string, format: uuid }
        roleCode: { type: string }
        orgId: { type: string, format: uuid, nullable: true }

    # ---- plans ----
    Plan:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        tier: { type: string }
        defaultTtlDays: { type: integer }
        active: { type: boolean }
        permissions:
          type: array
          items: { type: string }
        features:
          type: object
          additionalProperties: true
      required: [id, code, name]
    PlanCreate:
      type: object
      required: [code]
      properties:
        code: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 255 }
        description: { type: string }
        tier: { type: string, maxLength: 32 }
        defaultTtlDays: { type: integer }
        active: { type: boolean }
        permissions:
          type: array
          items: { type: string }
        features:
          type: object
          additionalProperties: true
    PlanUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        description: { type: string }
        tier: { type: string, maxLength: 32 }
        defaultTtlDays: { type: integer }
        active: { type: boolean }

    # ---- subscriptions ----
...
    SsoProvider:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        type:
          type: string
          enum: [SAML, OIDC]
        config:
          type: string
          description: Serialized JSON config (provider-specific).
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, orgId, type, enabled]
    SsoProviderCreate:
      type: object
      required: [type, config]
      properties:
        type:
          type: string
          enum: [SAML, OIDC]
        config:
          type: object
          additionalProperties: true
          description: Provider-specific config; see `docs/integration-guide.md`.
    SsoTestResult:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string }
        metadata:
          type: object
          additionalProperties: true

    # ---- apikeys ----
...